Move desktop / mobile filter to Filters (not tested)
